Remove MutableArrayMixin, ArrayMixin, NativeArrayMixin cycle
