+++ packages/ember-runtime/tests/mixins/enumerable_test.js
@@ -1,6 +1,6 @@
 import Enumerable from '../../mixins/enumerable';
 import ArrayProxy from '../../system/array_proxy';
-import { A } from '../../system/native_array';
+import { A } from '../../mixins/array';
 
 QUnit.module('Enumerable');
