+++ packages/ember-runtime/tests/mixins/mutable_array_test.js
@@ -1,8 +1,7 @@
 import { computed } from 'ember-metal';
 import MutableArrayTests from '../suites/mutable_array';
-import MutableArray from '../../mixins/mutable_array';
 import EmberObject from '../../system/object';
-import { A as emberA } from '../../system/native_array';
+import { A as emberA, MutableArray } from '../../mixins/array';
 import {
   arrayContentDidChange,
   arrayContentWillChange
