+++ packages/ember-runtime/tests/suites/array/any.js
@@ -1,5 +1,5 @@
 import { SuiteModuleBuilder } from '../suite';
-import { A as emberA } from '../../../system/native_array';
+import { A as emberA } from '../../../mixins/array';
 
 const suite = SuiteModuleBuilder.create();
