+++ packages/ember-runtime/tests/system/array_proxy/length_test.js
@@ -1,7 +1,7 @@
 import ArrayProxy from '../../../system/array_proxy';
 import EmberObject from '../../../system/object';
 import { observer, computed } from 'ember-metal';
-import { A as a } from '../../../system/native_array';
+import { A as a } from '../../../mixins/array';
 
 QUnit.module('Ember.ArrayProxy - content change (length)');
